requirements.txt
kagglehub==0.2.9
pandas==1.5.3
matplotlib==3.6.3
scikit-learn==1.2.2
kafka-python==2.0.2
pyspark==3.4.1
numpy==1.23.5
scipy==1.9.3
joblib==1.2.0
threadpoolctl==3.1.0
python-dateutil==2.8.2
pytz==2022.7
tzdata==2022.7
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
packaging==22.0
pillow==9.4.0
pyparsing==3.0.9
six==1.16.0